# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExecFleet API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/admin/sign-in": {
      "post": {
        "tags": [
          "Admin"
        ],
        "summary": "Admin sign-in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string"
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Admin signed in successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "token": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/admin/route": {
      "post": {
        "tags": [
          "Admin",
          "Routes"
        ],
        "summary": "Create a new route",
        "description": "Creates a new route and links it with a driver and optionally their vehicle.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Morning Route"
                  },
                  "routes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "employee_id": {
                          "type": "string",
                          "description": "MongoDB ObjectId of the employee",
                          "example": "64cabc1234567890abcdef12"
                        },
                        "address": {
                          "type": "string",
                          "example": "123 Main Street"
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number",
                              "example": 24.8607
                            },
                            "lng": {
                              "type": "number",
                              "example": 67.0011
                            }
                          }
                        }
                      },
                      "required": [
                        "employee_id",
                        "address",
                        "location"
                      ]
                    }
                  },
                  "city": {
                    "type": "string",
                    "example": "Karachi"
                  },
                  "from": {
                    "type": "string",
                    "example": "Office"
                  },
                  "to": {
                    "type": "string",
                    "example": "Client Site"
                  },
                  "kilometers": {
                    "type": "number",
                    "example": 12.5
                  },
                  "driver_id": {
                    "type": "string",
                    "description": "MongoDB ObjectId of the driver",
                    "example": "64df1234567890abcdef12"
                  }
                },
                "required": [
                  "name",
                  "routes",
                  "city",
                  "from",
                  "to",
                  "kilometers",
                  "driver_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string",
                      "example": "64cabcd9876543210fedcba1"
                    },
                    "name": {
                      "type": "string",
                      "example": "Morning Route"
                    },
                    "city": {
                      "type": "string",
                      "example": "Karachi"
                    },
                    "from": {
                      "type": "string",
                      "example": "Office"
                    },
                    "to": {
                      "type": "string",
                      "example": "Client Site"
                    },
                    "kilometers": {
                      "type": "number",
                      "example": 12.5
                    },
                    "driver_name": {
                      "type": "string",
                      "example": "Ali Raza"
                    },
                    "driver_phone": {
                      "type": "string",
                      "example": "03001234567"
                    },
                    "vehicle_number": {
                      "type": "string",
                      "example": "ABC-123"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid request body"
          },
          "404": {
            "description": "Driver or related resource not found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/admin/route": {
      "put": {
        "tags": [
          "Admin - Routes"
        ],
        "summary": "Update a route by ID",
        "parameters": [
          {
            "name": "route_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "MongoDB ObjectId of the route"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "routes": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "properties": {
                        "employee_id": {
                          "type": "string"
                        },
                        "address": {
                          "type": "string"
                        },
                        "location": {
                          "type": "object",
                          "properties": {
                            "lat": {
                              "type": "number"
                            },
                            "lng": {
                              "type": "number"
                            }
                          },
                          "required": [
                            "lat",
                            "lng"
                          ]
                        }
                      },
                      "required": [
                        "employee_id",
                        "address",
                        "location"
                      ]
                    }
                  },
                  "city": {
                    "type": "string"
                  },
                  "from": {
                    "type": "string"
                  },
                  "to": {
                    "type": "string"
                  },
                  "kilometers": {
                    "type": "number"
                  },
                  "driver_id": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "routes",
                  "city",
                  "from",
                  "to",
                  "kilometers",
                  "driver_id"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Route updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "routes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "employee_id": {
                            "type": "string"
                          },
                          "address": {
                            "type": "string"
                          },
                          "location": {
                            "type": "object",
                            "properties": {
                              "lat": {
                                "type": "number"
                              },
                              "lng": {
                                "type": "number"
                              }
                            }
                          }
                        }
                      }
                    },
                    "city": {
                      "type": "string"
                    },
                    "from": {
                      "type": "string"
                    },
                    "to": {
                      "type": "string"
                    },
                    "kilometers": {
                      "type": "number"
                    },
                    "driver_id": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request or validation failed"
          },
          "404": {
            "description": "Route not found"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}